Change logic for setting compliance flag
Previously, the logic used the presence or absence of comments as the
parameter to trigger whether or not an issue should be marked as
non-compliant or not.

It just so happened that before, the _only_ rule that added a comment
happened to be one that needed this behavior.

In 271f8a9, we added comments to two more rules - but these rules had
behavior that automatically brought the issue into compliance. There was
no need to set and then later unset the flag.

This change adds a new field to the context "non-compliant" that
defaults to False. A rule needs to explicitly set that to true in order
to request that the issue be updated with the flag.