Add check_rank automation for Enterprise Contract
Introduces a config file for EC. The main motivation is to include
the check_rank automation as suggested by KONFLUX-2392.

I'm not confident that we want the other available checks, (for
example we do have stories without epics, and epics without parent
features, and we do choose our own priority for bugs and stories
when grooming), so I'm just adding the `check_rank` option to begin
with.

However we don't do much ranking currently, so I think it's safe to
try letting the automation do some ranking for us.

(See config/template.yaml for the other automations available.)

Ref: https://issues.redhat.com/browse/EC-578
Ref: https://issues.redhat.com/browse/KONFLUX-2392