Don't fail if zero results found
This was added as a protection against misconfiguration, assuming that
if zero issues were found, then there must be a typo in the config.

It's actually a legitimate scenario though - we may have a project that
for a time will have zero issues open, but on which we don't want the
tool to fail hard.